Projects/Spotify historical from Billboard/requirements.txt
requests~=2.32.3
spotipy~=2.24.0
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1